Add HasMethod and IsInstanceOf to PPB_Var, and HasMethod to PPP_Class.
IsInstanceOf allows a plugin to unwrap objects that it created using
PPB_Var::CreateObject.  This allows a plugin that has its own object
system to maintain object identity.  It also helps with performance
if objects are not wrapped multiple times!

This patch also adds support to Var for ScriptableObject, and the
example plugin is revised to make use of ScriptableObject.

R=brettw
BUG=none
TEST=none

Review URL: http://codereview.chromium.org/2129017