Add a test for stream-to-file.  Hook up URLResponseInfo::GetBody.
Remove some unimplemented security properties from URLRequestInfo.
I'm not convinced that these are the flags that we want, so it is
best to remove them for now.

Add a FinishStreamingToFile method to PPB_URLLoader.

Change the documentation for PPB_URLResponseInfo::GetBody to
indicate that the file reference remains valid for the lifetime
of the *URLLoader* instead of the *URLResponseInfo*.  This makes
sense given that an URLResponseInfo object is just meant to be
a container for data, whereas the URLLoader corresponds to the
act of loading an URL.  (This change also matches the WebKit API
for downloading to a file, which URLLoader is implemented on top
of.)

R=brettw
BUG=49789
TEST=TestURLLoader.TestStreamToFile

Review URL: http://codereview.chromium.org/3040024