Fix flakiness by running tests asynchronously.
I considered adding an explicit JS call for this, but that would require
overriding TestingInstance::GetInstanceObject, which already has meaning for
some tests.  It might be odd to return a different object from that function
depending on what state we are in.

I had to modify test_case.html to receive an explicit callback from the plugin
when execution of tests conclude.  This is used to know when we should resize
the IFRAMEs used to embed tests in "RunAll" mode.  This is actually a lot less
hacky then what I was doing before.

At any rate, this solution is fairly straightforward.

R=brettw

Review URL: http://codereview.chromium.org/3323015