Add module parameters to the necessary interfaces so we can associate strings and object vars with modules internal to the browser.
This adds a new testing interface to count the number of open objects which
allows us to write tests to verify there were no leaks. With this I found a
leak in the C++ wrapper handling output exceptions which this fixes.

I removed the Var swap operator for Var. This was confusing me when trying to
debug the leak, and I'm planning on removing all of the swap functions since
they're actually not more efficient than std::swap for our case.

TEST=included
BUG=http://code.google.com/p/chromium/issues/detail?id=43964
Review URL: http://codereview.chromium.org/3341017