Revert PP_Bool change.
We try to submit a Chrome change to pull PPAPI deps and fix the build ASAP
after landing a PPAPI change, and the changes should be tested on the trybot
together using the -s option as described on
http://www.chromium.org/developers/design-documents/pepper-plugin-implementation

I couldn't find you online and it's Friday evening so I doubt this will be
fixed before the weekend. In the meantime, all PPAPI committers which pull the
writable repo on trunk are broken and need to edit their .gclient, and nobody
else can submit any changes to PPAPI. Reverting for now to unblock.

This infrastructure problem should be fixed when I move PPAPI to the Chrome
repo.
Review URL: http://codereview.chromium.org/4149008